chore: Fix clippy warnings across codebase
- Replace clone() with dereference for Copy types in bind.rs
- Use function reference instead of closure in storage.rs
- Replace manual div_ceil with len.div_ceil(64) in simd_ops
- Use .cloned() instead of .map(|r| r.clone()) in fast_path.rs
- Convert match to if let for single-pattern matches
- Make validate_expression_columns a static method
- Use .flatten() instead of nested if let Some loops
- Collapse nested if let patterns in sysbench_benchmark.rs
- Remove unnecessary type casts in tpce benchmarks
- Use iter().take(n) instead of indexed loop
- Replace expect(&format!()) with unwrap_or_else(|_| panic!())
- Use Option::map instead of if let Some/else None
- Replace len() >= 1 with !is_empty()
- Remove redundant local variable bindings
- Use .first() instead of .get(0)
- Collapse nested if statements with &&

+++ crates/vibesql-executor/benches/tpce/data.rs
@@ -923,7 +923,7 @@ impl TPCEData {
             s_name: format!("{} Common Stock", s_symb),
             s_ex_id: self.rng.pick(&exchanges).to_string(),
             s_co_id: co_id,
-            s_num_out: self.rng.random_int(1_000_000, 1_000_000_000) as i64,
+            s_num_out: self.rng.random_int(1_000_000, 1_000_000_000),
             s_start_date: self.rng.random_date(1980, 2020),
             s_exch_date: self.rng.random_date(1980, 2020),
             s_pe: self.rng.random_decimal(5.0, 50.0, 2),
@@ -944,7 +944,7 @@ impl TPCEData {
             lt_dts: TPCERng::current_timestamp(),
             lt_price: price,
             lt_open_price: price * self.rng.random_float(0.95, 1.05),
-            lt_vol: self.rng.random_int(1000, 10_000_000) as i64,
+            lt_vol: self.rng.random_int(1000, 10_000_000),
         }
     }
 
@@ -953,7 +953,7 @@ impl TPCEData {
         let revenue = self.rng.random_decimal(1_000_000.0, 1_000_000_000.0, 2);
         let margin = self.rng.random_decimal(0.05, 0.30, 4);
         let net_earn = revenue * margin;
-        let shares = self.rng.random_int(10_000_000, 1_000_000_000) as i64;
+        let shares = self.rng.random_int(10_000_000, 1_000_000_000);
 
         Financial {
             fi_co_id: co_id,
